pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-auth>=2.45.0",
    "gspread>=6.2.1",
    "oauth2client>=4.1.3",
    "pandas>=2.3.3",
    "python-levenshtein>=0.27.3",
    "requests>=2.32.5",
    "streamlit>=1.52.2",
    "thefuzz>=0.22.1",
]
